Move utils to pkg/utils
