+++ commands/peers/deletepeer.go
@@ -5,9 +5,9 @@ import (
 
 	"github.com/gluster/glusterd2/gdctx"
 	"github.com/gluster/glusterd2/peer"
+	"github.com/gluster/glusterd2/pkg/utils"
 	restutils "github.com/gluster/glusterd2/servers/rest/utils"
 	"github.com/gluster/glusterd2/store"
-	"github.com/gluster/glusterd2/utils"
 	"github.com/gluster/glusterd2/volume"
 
 	"github.com/gorilla/mux"
