+++ commands/volumes/common.go
@@ -5,9 +5,9 @@ import (
 	"strings"
 
 	"github.com/gluster/glusterd2/gdctx"
+	"github.com/gluster/glusterd2/pkg/utils"
 	"github.com/gluster/glusterd2/servers/sunrpc"
 	"github.com/gluster/glusterd2/transaction"
-	"github.com/gluster/glusterd2/utils"
 	"github.com/gluster/glusterd2/volgen"
 	"github.com/gluster/glusterd2/volume"
 	"github.com/gluster/glusterd2/xlator"
