+++ gdctx/global.go
@@ -12,7 +12,7 @@ import (
 	"os"
 	"path"
 
-	"github.com/gluster/glusterd2/utils"
+	"github.com/gluster/glusterd2/pkg/utils"
 	"github.com/gluster/glusterd2/version"
 
 	"github.com/pborman/uuid"
