+++ volgen/cluster-graph.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 
 	"github.com/gluster/glusterd2/brick"
-	"github.com/gluster/glusterd2/utils"
+	"github.com/gluster/glusterd2/pkg/utils"
 	"github.com/gluster/glusterd2/volume"
 )
