+++ volgen/volgen.go
@@ -9,8 +9,8 @@ import (
 	"strings"
 
 	"github.com/gluster/glusterd2/brick"
+	"github.com/gluster/glusterd2/pkg/utils"
 	"github.com/gluster/glusterd2/store"
-	"github.com/gluster/glusterd2/utils"
 	"github.com/gluster/glusterd2/volume"
 )
