+++ volume/struct.go
@@ -12,7 +12,7 @@ import (
 	"github.com/gluster/glusterd2/brick"
 	"github.com/gluster/glusterd2/gdctx"
 	"github.com/gluster/glusterd2/peer"
-	"github.com/gluster/glusterd2/utils"
+	"github.com/gluster/glusterd2/pkg/utils"
 
 	"github.com/pborman/uuid"
 	log "github.com/sirupsen/logrus"
